pyproject.toml
[project]
name = "git-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastmcp>=2.2.2",
    "gitpython>=3.1.44",
]

[dependency-groups]
dev = [
    "ruff>=0.11.6",
]
test = [
    "pytest-dotenv>=0.5.2",
]
